+++ R/autosize.R
@@ -55,19 +55,36 @@ vw_autosize <- function(spec, width = NULL, height = NULL) {
     call. = FALSE)
 }
 
+.autosize.vegaspec_hconcat <- function(spec, width = NULL, height = NULL) {
+  .warn_autosize(class(spec))
+  NextMethod()
+}
+
+.autosize.vegaspec_vconcat <- function(spec, width = NULL, height = NULL) {
+  .warn_autosize(class(spec))
+  NextMethod()
+}
+
+.autosize.vegaspec_concat <- function(spec, width = NULL, height = NULL) {
+  .warn_autosize(class(spec))
+  NextMethod()
+}
+
+.autosize.vegaspec_facet <- function(spec, width = NULL, height = NULL) {
+  .warn_autosize(class(spec))
+  NextMethod()
+}
+
+.autosize.vegaspec_repeat <- function(spec, width = NULL, height = NULL) {
+  .warn_autosize(class(spec))
+  NextMethod()
+}
+
 .autosize.vegaspec_vega_lite <- function(spec, width = NULL, height = NULL) {
 
   if (is.null(c(width, height))) {
     # nothing to do here
     return(spec)
-  }
-
-  # if this spec has multiple views, warn that autosize will not work
-  if (is_multiple_view(spec)) {
-    warning(
-      "Specifying the width or height of a ",
-      "vegaspec with multiple views has no effect on rendering."
-    )
   }
 
   # using this notation: spec$config <- spec$config %||% list()
@@ -108,11 +125,13 @@ vw_autosize <- function(spec, width = NULL, height = NULL) {
   spec
 }
 
-is_multiple_view <- function(spec) {
-
-  spec <- as_vegaspec(spec)
-
-  names_multiple_view <- c("facet", "repeat", "hconcat", "vconcat")
-
-  any(names_multiple_view %in% names(spec))
+.warn_autosize <- function(class) {
+  warning(
+    glue::glue(
+      "This vegaspec has class `{class[[1]]}`, which implies multiple views. ",
+      "Specifying the width or height of a ",
+      "vegaspec with multiple views has no effect on rendering."
+    ),
+    call. = FALSE
+  )
 }