+++ NEWS.md
@@ -1,5 +1,7 @@
 # vegawidget (0.2.0 development version)
 
+* Move **shiny** from an `Imports` dependency to a `Suggests` dependency (#83). 
+
 * Modify `knit_print()` to handle non-HTML formats. Instead of using the [webshot](https://github.com/wch/webshot) package, the `knit_print()` function converts charts to either `png`, `svg` (useful for `github_document`), or `pdf` (useful for `pdf_document`) format. (#44, @AliciaSchep)
 
 * (Of interest to developers) modify the class-naming for Vega-Lite specs, adding a another level of S3 class that describes the type of chart, e.g. unit, layer, facet, etc. (#77)